Makes peace between Mimemail and Rules.
When installing with Drush's site-install, Rules would fire
ReflectionExceptions. Also, that would happen sometimes and randomly inside
the website, rendering it unusable.

The reason for this is Rules actions file being in a module's sub-directory.
Couple that with module_load_include() trying to do

    drupal_get_path('module', 'subdir/my.file')

and you get what we got.

More information about this here:

* http://drupal.org/node/1235956
* http://drupal.org/node/1585546

Change-Id: I5b4e5961b29d5254f91c14c4c057a9242a4367ae
Reviewed-on: https://code.commerceguys.com/3728
Reviewed-by: Hudson bot <bender@commerceguys.com>
Reviewed-by: Bojan Zivanovic <bojan@commerceguys.com>