Fixes features roulette. See #8720.
Features caches files it includes in one request. Basically, it tries
to do reqiure_once but on it's own, instead to let that job to
require_once which is faster.

And it caches wrong data: a feature component. It works fine until you
try to install/rebuild several features in one go, as we do in Drush.

This patch lets PHP worry to require only once.

Change-Id: Ie28304329083d3db9be086ad77bfc7b7290ec26d
Reviewed-on: https://code.commerceguys.com/3785
Reviewed-by: Hudson bot <bender@commerceguys.com>
Reviewed-by: Bojan Zivanovic <bojan@commerceguys.com>