Fixed breadcrumbs.
The state of breadcrumb manipulation in D7 is not so good. Hence, we
have to use two modules for the breadcrumbs:

* taxonomy_breadcrumb
  For showing the taxonomy crumbs, ie. all of our products.

* custom_breadcrumbs (dev version)
  We need to use the dev version (2.x branch) since in it is support
  for views breadcrumbs. Other than the views, we use it for simple
  content types, blog and page ones.

This commit includes two patches:

1) taxonomy_breadcrumb-view_mode_bug-1280434-2.patch
This is a fix for the '=' vs. '==' error in a module.

2) taxonomy_term-1608070-1_views_support.patch
Adds views support.

Taxonomy breadcrumb module is poorly supported with a couple of issues
about project takeover.

Custom breadcrumbs do not sport features support, so in the blog_ui we
are creating breadcrumbs directly. Not the pretties solution,
but... The rest of custom_breadcrumbs variables are exported into
blog_ui via Strongarm.

Change-Id: I736591f954c879548112d253e19f0d5873a7efe5
Reviewed-on: https://code.commerceguys.com/3833
Reviewed-by: Bojan Zivanovic <bojan@commerceguys.com>