Re-order dependencies for SimpleTest
DrupalWebTestCase::setUp activates a profile's dependencies via
module_enable($profile_details['dependencies'], FALSE);

The "FALSE" means that only the specified modules will be enabled, and
only in the specified order. Since $profile_details['dependencies']
comes straight from the profile's .info file, that file needs to be
complete and in a valid order.

This change ensures that.